# docs/swagger/swagger.yaml
definitions:
...
  model.TagDTO:
    properties:
      bookmark_count:
        description: Number of bookmarks with this tag
        type: integer
      deleted:
        description: Marks when a tag is deleted from a bookmark
        type: boolean
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/tags:
    get:
      description: List all tags
      parameters:
      - description: Include bookmark count for each tag
        in: query
        name: with_bookmark_count
        type: boolean
      - description: Filter tags by bookmark ID
        in: query
        name: bookmark_id
        type: integer
      - description: Search tags by name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.TagDTO'
            type: array
        "403":
          description: Authentication required
        "500":
          description: Internal server error
      summary: List tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      description: Create a new tag
      parameters:
      - description: Tag data
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/model.TagDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.TagDTO'
        "400":
          description: Invalid request
        "403":
          description: Authentication required
        "500":
          description: Internal server error
      summary: Create tag
      tags:
      - Tags
...
swagger: "2.0"